fix job_tracker and async_task Python logging
if PYARROW_IGNORE_TIMEZONE is not set, it will output this and suppress all other logging:

WARNING:root:'PYARROW_IGNORE_TIMEZONE' environment variable was not set. It is required to set this environment variable to '1' in both driver and executor sides if you use pyarrow>=2.0.0. pandas-on-Spark will set it for you but it does not work if there is a Spark context already launched.