add poc for persistent workers
This POC is locked behind an environment variable. It will initially only be used by the openeo-integrationtests-poc Jenkins pipeline. The idea is to run smaller batch jobs on a number of pre-started containers that run until the main driver shuts down. This will lower the amount of time users spent waiting on queued jobs significantly. The final product should use a message queue such as Apache Kafka but currently the shared filesystem is used for communication.