#302 use simple error ratio check in orfeo executor
Getting the value of an accumulator is not allowed in pyspark, while it is allowed in spark. To get around this we start with a simple implementation without accumulators. Eventually we can use a singleton for each batch job process that checks at the very end of batch_job.py if the soft error ratio has been exceeded and then notifies the user.