bugfix: no need to serialize job results dependencies #489
We schedule batch processes ourselves, then pass on their source_location (and card4l flag) to the batch job that is subsequently started upon their completion ("DONE"). The load_collections that get invoked while evaluating the process graph in the batch job only rely on source_location and card4l, not on the arguments actually passed to load_collection.

In the case of unfinished job results dependencies, however, there's no need to serialize them and pass them on to the subsequent batch job, because load_stacs evaluated in that batch job will simply act upon the very same "running" URL that made it a dependency in the first place (but this time it will actually read the job results because they have "finished" in the meanwhile).