feature: provide s3 profiles and tokens during execution
When a job gets submitted generate an identity token which can be exchanged for AWS credentials that can be used against the S3 proxy but scoped to the job context.

Additionally a config file gets generated with AWS Profiles which can then be used in order
to access S3 via the S3 proxy. This way using an AWS profile is the only config needed for your S3 access. The profile names correspond to workspace names for workspace access.

For normal usage we add a:
 - base profile which overtime can be used for access to OpenEO platform resources (e.g. results bucket, logs, ...)
 - eodata profile which allows access to the EOdata collection.

This change is behind a config feature flag which defaults to false to allow gradual rollout and testing.