PR-feedback file watcher
The file watcher depended on global state. If the situation occurs that 2 components use the same config file it could be that 1 component observes the change and marks it as treated and the other component will not know about the change.

By introducing a watcher class with its own state every component can have its private bookkeeping.