tests: disable proxy tests
Currently these tests fail with 5XX errors. This seems to be an issue on the CICD enviornment where it is likely that moto does not come up correctly and the requests gets sent to a wrong process.

Given that this is not yet a production feature and shielded with config flags we disable the tests until they are reliable