credentials: remove provider from regions specific variable names
Given that the assumption is anyway that region names are unique enough such that provider-region_name is a 1-M relationship it is better to not have the provider name in the environment variable of the region-specific credentials. That way the driver does not need to know the provider in order to still work with that particular storage provider.