Workspace/allow cross region (#1192)
* workspaces: actually use region specific config

Workspaces track a region but client config just takes environment variable which assume workspace is in the same region as the execution environment its object storage.

This change adds openeogeotrellis.integrations.s3.get_s3_client which takes a region and then configures the client as required.

* refactor: small move to break dependency cycle

Will still review better option but unblock tests

* Support unknown providers.

This makes the behavior closer to before support of multiple regions
where fallback is done to a specific configuration in a specific
environment variable (SWIFT_URL)

* credentials: remove provider from regions specific variable names

Given that the assumption is anyway that region names are unique enough such that provider-region_name is a 1-M relationship it is better to not have the provider name in the environment variable of the region-specific credentials. That way the driver does not need to know the provider in order to still work with that particular storage provider.

* tests: have default SWIFT_URL

Workspace tests mock away interaction but also did not handle minimum config requirements.

* refactor: use S3ClientBuilder and CustomStacIO regionaware)

* tests: fix after refactor

* tests: mock s3 client when boostrapping s3 resources

When mocking s3 resources one expects access to S3 to be mocked as well.

* refactor: use local import

* Empty integrations/s3/__init__.py to break (future) import cycles

related to #1190/#1192

* refactor: local import no longer needed after #1194

* pr-feedback: naming fixture and explain why explicit monkey patching

* doc: CHANGELOG.md update

---------

Co-authored-by: Peter Van Bouwel <peter.vanbouwel@vito.be>
Co-authored-by: Stefaan Lippens <stefaan.lippens@vito.be>