save randomforest model to fs mounts instead of s3a eu-cdse/openeo-cdse-infra#636
The s3a implementation aborts the multipart upload of the randomforest model from the executors to the batch job's bucket. Given that we now properly mount these buckets on our kubernetes deployments on both the driver and executors, we can now use simply save these results to the filesystem. These buckets are mounted via s3fs so we still have to be careful about eventual consistency issues where we archive and remove the model directory before all results are present. However, all tests so far indicate that this is no problem.