fix loading catboost model in modelloader.py
After testing I can see that `CatBoostClassificationModel.loadNativeModel()` can just use a path that is local to the driver. This simplifies the implementation. There was also a bug where it was passing a PosixPath instead of a string to the java implementation, and it was not using the actual model file but its parent directory.