fix: run_as_group as root
In the new calrissian container the Python binary is on a path owned by the calrissian user and root group. Since the calrissian user has id 1001 rather than 1000 we should make sure we use a group that matches the calrissian binary and Python runtime binaries (group 0). Changing the run_as_user to 1001 will create problems with the cloudinsar image because there volumes will only be accessible by a user with id 1000 and calrissian does not support fs_group yet.