skip test_load_stac_from_spatiotemporal_netcdf_mixed_columns
For some reason this crashes the JVM when `--add-opens=java.base/sun.net.util=ALL-UNNAMED` is added to `spark.driver.extraJavaOptions`. The hypothesis is that this has something to do with multithreading, where multiple threads try to access the same memory. However it is strange that this java option activates the issue.

ref https://github.com/eu-cdse/openeo-cdse-infra/issues/690